Изменения по комментариям
